Revert refactor of flag_output
It was wrong to place it ahead of the while loop, because some filtering
flags affect *how* output is shown, but not whether it is/isn't, so we
can't deduce that we'll output anything based on these alone.

e.g.
    note -d -s
    note -s -d
should print short forms of today's notes, but
    note -s
should print nothing.